Improve formatting of api route specs
