+++ spec/routing/api_routing_spec.rb
@@ -1,30 +1,49 @@
 require 'rails_helper'
 
 describe 'API routes' do
-  describe 'credentials routes' do
-    it 'routes verify credentials' do
+  describe 'Credentials routes' do
+    it 'routes to verify credentials' do
       expect(get('/api/v1/accounts/verify_credentials')).
         to route_to('api/v1/accounts/credentials#show')
     end
 
-    it 'routes update credentials' do
+    it 'routes to update credentials' do
       expect(patch('/api/v1/accounts/update_credentials')).
         to route_to('api/v1/accounts/credentials#update')
     end
+  end
 
-    it 'routes account statuses' do
+  describe 'Account routes' do
+    it 'routes to statuses' do
       expect(get('/api/v1/accounts/user/statuses')).
         to route_to('api/v1/accounts/statuses#index', account_id: 'user')
     end
 
-    it 'routes account search' do
+    it 'routes to search' do
       expect(get('/api/v1/accounts/search')).
         to route_to('api/v1/accounts/search#show')
     end
 
-    it 'routes account relationships' do
+    it 'routes to relationships' do
       expect(get('/api/v1/accounts/relationships')).
         to route_to('api/v1/accounts/relationships#index')
+    end
+  end
+
+  describe 'Timeline routes' do
+    it 'routes to home timeline' do
+      expect(get('/api/v1/timelines/home')).
+        to route_to('api/v1/timelines/home#show')
+    end
+
+    it 'routes to public timeline' do
+      expect(get('/api/v1/timelines/public')).
+        to route_to('api/v1/timelines/public#show')
+    end
+
+    it 'routes to tag timeline' do
+      expect(get('/api/v1/timelines/tag/test')).
+        to route_to('api/v1/timelines/tag#show', id: 'test')
     end
   end
 end
